RISC-V: Support instruction prefixes as opcode names
Look up opcodes using progressively shorter prefixes of the instruction
name stopping at a period.  This condenses the opcode tables for Hwacha
by allowing the vector/scalar suffix to be treated as an argument.

Since RVV and Hwacha overlap in some mnemonics, the initial opcode entry
to be found might not match in the instruction class, so the main
assembler loop should be re-executed after each attempt.